requirements.txt
fastapi
requests
python-dotenv
uvicorn
pandas